# Improve ackManager.completeTask performance by two orders of magnitude