@dnr 2023-12-08
Oh, I see, you picked that one since it's already used in the test. Well, I'm still curious. If it's faster maybe we can change both