@tdeebswihart 2023-12-08
> Not to be too pedantic since it's already a huge improvement, but I'm curious how the performance would compare to using https://pkg.go.dev/github.com/google/btree, specifically the ordered generic version (NewOrderedG). That's the ordered map for Go I've used a few times in the past

Looks like this doesn't allow us to store values alongside the keys, hrm. I'll build it with this alongside a map of K->V to seet how it performs