@dnr 2023-12-08
> Looks like this doesn't allow us to store values alongside the keys, hrm. I'll build it with this alongside a map of K->V to see how it performs

Oh right, I forgot you need the bool. You can totally store values next to the keys, but yeah, you can't use the Ordered one in that case, you need to make a struct with two members and implement a Less